+++ forms/content-about-whatsnew.php
@@ -27,9 +27,9 @@ include( COREACTIVITY_PATH . 'forms/content-about-minor.php' );
 				<?php } ?>
 
                 <div class="coreactivity-about-counters">
-                    <div><i class="d4p-icon d4p-ui-folder d4p-icon-fw"></i> <strong>25</strong> Components</div>
-                    <div><i class="d4p-icon d4p-ui-radar d4p-icon-fw"></i> <strong>123</strong> Events</div>
-                    <div><i class="d4p-icon d4p-ui-plug d4p-icon-fw"></i> <strong>11</strong> Plugins</div>
+                    <div><i class="d4p-icon d4p-ui-folder d4p-icon-fw"></i> <strong>26</strong> Components</div>
+                    <div><i class="d4p-icon d4p-ui-radar d4p-icon-fw"></i> <strong>127</strong> Events</div>
+                    <div><i class="d4p-icon d4p-ui-plug d4p-icon-fw"></i> <strong>12</strong> Plugins</div>
                 </div>
             </div>
         </div>